requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers>=4.44.0
datasets>=2.18.0
peft>=0.12.0
bitsandbytes>=0.43.0
trl>=0.9.0
accelerate>=0.33.0
scipy>=1.11.0
pyyaml>=6.0.1
tqdm>=4.66.0
evaluate>=0.4.2
rouge_score>=0.1.2
bert_score>=0.3.13
openai>=1.40.0
matplotlib>=3.8.0
